Move Auth0RedirectRuleUtilities to own file
